- Fix integer truncation issue (int -> size_t).
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@32753 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/writers/writedb/writedb_impl.cpp
@@ -297,7 +297,7 @@ CWriteDB_Impl::x_BuildDeflinesFromBioseq(const CBioseq                  & bioseq
         defline->SetSeqid() = ids;
         ids.clear();
         
-        unsigned pos = titles.find(" >");
+        size_t pos = titles.find(" >");
         string T;
         
         if (pos != titles.npos) {
